config: rename options to show_revision_number and show_sha_length
'sha_show_numeric_rev' had nothing to do with the sha value.

The revision numbers are kind of native to Mercurial and there they are known
as 'revision numbers'.

'sha_len' was very short and didn't clarify that it only controlled what was shown.

These settings are currently only used in the changelog, but they should be used everywhere.