changelog: fix display artifacts in expanded multi-line commit messages
When a multi-line commit message is expanded in a changelog, and the last line
of the commit message has some characters that go below the baseline, like an
underscore, g, j, y, ... the bottom part of these characters would not be
shown.

This is caused by the 'overflow: hidden' property set on the unexpanded
message. Reset that property by adding 'overflow: initial' on the expanded
class.

Additionally, slightly enlarge the margin of the expanded message box.