pullrequests: add PullRequest.nice_id method
Currently, a pull request id is referenced as #5, and the '#' symbol is
fixed and repeated in several places. This commit adds a class method
make_nice_id, that returns a string reference to a pull request,
currently in the form '#5'.

This function could be overridden by an organization if they need references
in another form, for example PR-5.

The method nice_id uses this for the pullrequest's own id.