autocomplete: remove redundant specification of responseSchema
YAHOO.util.DataSource.responseSchema determines which fields from the data
source are returned upon requests. In the case of user/group autocompletion,
the data source is a static JavaScript array containing only the relevant
fields. When not specifying a responseSchema, all fields of the selected
entry are returned, which is fine (in fact, the responseSchema currently
specified just mentions all fields, redundantly).