cleanup: remove dead code and templates related to 'Switch To'
a33448d81f70 introduced a new 'Switch To' functionality and removed the last
reference to branch_tag_switcher and thus to switch_to_list.html and the tags,
bookmarks and branches pages.

We don't need these pages ... and if we do, we will implement them differently.
They are mostly YUI datatables with compare links that only can compare within
the same kind of revision names.