db: remove SQLAlchemy Migrate database table
Use Alembic to migrate database away from SQLAlchemy Migrate. This
eliminates the last vestiges of SQLAlchemy Migrate. Since we drop
the Migrate table (and its contents), it is not possible to revert
this database change; however, downgrading to this Alembic schema
revision (9358dc3d6828) will in practice be enough to ensure
compatibility with all previous Kallithea versions.

(As always, the Alembic migration script is committed in the same
revision as the database schema changes.)