db: drop length limits on TEXT columns
As of the upgrade to version 1.0 in 3c4b6ddf6735, SQLAlchemy began to
pass TEXT length limits to the database during table creation. Such
limits are however not supported by SQLite, MySQL nor PostgreSQL, and
while SQLite simply ignores it, it is a syntax error in PostgreSQL,
breaking the creation of new PostgreSQL databases. With the lengths
being unused and quite arbitrary, just drop them.