templates: fix dropdown functionality of current menu item
After 3e46e6b6a27a, the 'current' menu would have two class attributes and it
would not work as a dropdown. For example, it was not possible to navigate from
Admin>Users to Admin>User Groups.

Instead, get rid of the is_current "helper" and just handle classes explicitly.
That might be slightly more verbose but is also more explicit ... and it works.