css: fix redundant/overly broad rules
Remove rule coloring "#content div.panel div.panel-heading .pull-left a"
(and ditto .pull-right) white, as it caused the branch filter drop-down
box on the changelog page to become white-on-white.

I have not found anywhere this rule actually serves a purpose.

On several pages, the rules matched our standard grey buttons, but was
overruled by an !important style for .btn elements.

On repository group pages, it matches the repo group parent path link,
and in /_admin/journal, it matches the "Watched Repositories" and "My
Repositories" links; but all of these are already white from the
"#content div.panel div.panel-heading a" rule.