api: change precision of ChangesetStatus.modified_at to seconds
Per default MySQL only uses seconds. So in order to be consistent on all
databases, this is the easiest solution. I don't think the microseconds are
necessary. And AFAICS mercurial only uses seconds for the changeset
modification time as well. So why should we use microseconds for ChangesetStatus.

Without this, for example test_api_get_changeset_with_reviews fails, because of
datetime mismatch if MySQL is used.