celery: replace loader.PylonsSettingsProxy with a simple configuration object
The proxy was quite complex and did all the conversion at runtime, making it
hard to figure out exactly what it was doing.

Instead, just create a simple configuration object as described on
http://docs.celeryproject.org/en/3.1/userguide/application.html and convert all
relevant config settings up front. We don't convert all config settings, but
only the ones Celery seems to be interested in. The value conversion heuristics
are the same as before and still seem quite fragile.