style: use npm less to generate kallithea stylesheet style.css from style.less
This is a minimal change to introduce less as framework to generate our
stylesheet. The input less file is currently the same as the original css
stylesheet, so that the delta between the input style.less and the output
style.css is zero.

To regenerate style.css from style.less, run:

  npm --prefix kallithea/public/less install
  npm --prefix kallithea/public/less run less

These changes are based on work by Dominik Ruf.