diffs: do check of cut_off_limit on the raw diff before any parsing and allocation
This redefines the exact meaning of cut_off_limit but it will still be
approximately the same thing. It makes it more a limitation of what amount of
work should be done, more than how much html should be outputted.

It could make sense to push the limit further back to vcs to also prevent
computing or allocating memory for huge diffs.