changelog: bring the right-aligned tags in a least to most common order
The goal is, to more often, have equal tags being display like a column, without
giving each tag its own table cell, which would waste a lot of space.
So move tags and branch, which come in singles move in front of phase, unstable,
etc. which most of the time come in groups.