vcs: add support for Mercurial annotate in Mercurial 4.4
Mercurial fctx.annotate() returns tuples with two elements: info about the
annotate line and the actual line. The code is changed to clarify that.

After Mercurial 4.4 with https://www.mercurial-scm.org/repo/hg/rev/2e32c6a31cc7 ,
the info is no longer a tuple but an attr object. Assume fctx is available as
an attribute, but catch exceptions and fall back to indexing as before.

That can't easily be done in hgcompat, so we do it inline.