less: clean up changelog style
- use table-layout: fixed
Since we already use a fixed width for the most cells, this makes it easier
(faster) for the browser to render specially big tables.
And it allows to use 'text-overflow: ellipsis'. The old ellipsis settings
have no effect.

- use nowrap and 'text-overflow: ellipsis' for all cells
A table is easier (faster) to read for a user if all the row have the same height.
Therefore don't wrap lines and in case of overflow use ellipsis.

- remove unnecessary height
Up until now the row was bigger then 31px so this had no effect.
And it look wrong to make the row bigger then its content.

- remove unnecessary border-color
bootstrap default is almost the same color

- remove unnecessary text-align
left is the default

- remove font-size or use bootstrap .small
Making fonts too small hurts the readability.

- remove some padding and margin

- don't use special colors
The differences are hardly noticable or in case of the gray dates hurt readability.

- remove obsolete table#changesets .logtags style