MANIFEST: only include bootstrap from node_modules when packaging or installing
With the 'less' source folder placed under 'public', the whole folder happened
to be fully included in source distribution and installed. *If* 'npm install'
has been run, 'kallithea/public/less/node_modules/' will contain both
build/development tools ('less' and dependencies) and the source of Bootstrap.

We do want Bootstrap source to be included so it automatically gets 'vendored',
but we don't want all the build tools.

We thus prune the whole 'node_modules' directory, but add
'node_modules/bootstrap' back in.

This is exactly what it is. It might seem a bit fragile: There is no
verification that it only is the 'bootstrap' directory that should be added
back in, and there is no verification that it actually is present and included.