setup-db: print completion message to avoid confusion (issue #303)
There are cases where the last message of setup-db is a warning, giving the
impression that the overall command failed which may not actually be the
case.

For example, when git is not installed, warnings are given, but they are not
an actual error. Kallithea will work fine for Mercurial repositories.

To avoid any confusion, print a completion message at the end.
Any real errors will abort the command and not make it this far.