HTML email templates: Outlook: center action button
The rendering of HTML emails in Outlook is very restricted and based on old
Internet Explorer versions (more or less the same rendering as in Microsoft
Word). To make the Kallithea emails look the same as in the browser, some
adaptations are needed.

One of the problems is that the action button is not centered.
An extra <center> tag solves this issue.