HTML email templates: Outlook: fix padding inside of action button
The rendering of HTML emails in Outlook is very restricted and based on old
Internet Explorer versions (more or less the same rendering as in Microsoft
Word). To make the Kallithea emails look the same as in the browser, some
adaptations are needed.

The padding inside the button is completely lost in Outlook, the text sticks
to the border. This can be fixed by moving the padding styling from the
<center> tag to the <td> cell.