HTML email templates: Outlook: fix width and height attributes to preserve whitespace
The rendering of HTML emails in Outlook is very restricted and based on old
Internet Explorer versions (more or less the same rendering as in Microsoft
Word). To make the Kallithea emails look the same as in the browser, some
adaptations are needed.

HTML width and height attributes do not have any effect, the size is
interpreted as 0px. A CSS-based width and height do work in Outlook and
result in the expected spacing between and within elements of the email.

For Outlook rendering, the CSS variant is probably sufficient, but as it is
unclear how other mail agents will interpret this, leave the HTML width and
height attributes as well.