HTML email templates: Outlook: fix list bullets
The rendering of HTML emails in Outlook is very restricted and based on old
Internet Explorer versions (more or less the same rendering as in Microsoft
Word). To make the Kallithea emails look the same as in the browser, some
adaptations are needed.

List bullets (to identify different changesets in pullrequests) are not
shown in Outlook. In order to fix that, a proprietary CSS property is
required.
Fix found via:
https://litmus.com/community/discussions/752-can-we-use-ul-and-li-in-email