HTML email templates: don't use link color for PR descriptions
The link color should only be used for actual hyperlinks.
Using it for anything else is very confusing for users.

For pullrequest mails, the usage of the link color for the PR description
may have been done to separate the description more clearly from the title
word 'Description' above it.
Instead, use a table with grey background similar to how comments are
displayed. This meets the objective of not abusing link colors as well as
clearly separating the description from the rest of the mail.