email templates: reorder references to pullrequests
The current references to PRs in email templates are:

    Pull request from <repo> branch <branch> to <repo> branch <branch>:
    <niceid> <title> by <user>

as the first line can become very long and therefore hard to read, a first
step is to introduce a newline before the destination branch:

    Pull request from <repo> branch <branch>
    to <repo> branch <branch>:
    <niceid> <title> by <user>

But it would be nicer if the from/to part aligns more.
Moreover, the most useful part is (hopefully) the PR title, the from/to is
secondary and can sometimes even be inferred from the title and context the
recipient is already aware of.

Thus, this commit proposes the following format:

    Pull request <niceid> <title> by <user>
    from <repo> branch <branch>
    to <repo> branch <branch>


This adds some divergence with the changeset references, which could be
aligned later if desired.