diff: fix 2-way diff panes not being shown side by side
eee7674cef41 had the side effect of setting overflow styling on the DOM
elements used by mergely. That confused it's computation of sizes on some
browsers, scroll bars got added, and the panes were re-flowed to not be
side-by-side.

To fix that, add an exception that allow overflow-x to be set to visible.