graph: make sure graph is shown next to the table it belongs to
For some screens sizes (for example around 800 px), the graph with changes
available for PR update wasn't shown next to the corresponding table. The two
boxes would float vertically and be rendered after each other.

To fix that, use the same trick as before 02aef1484695: set the graph height to
0 and give the table a left margin leaving room for the graph.

(The fancyness of Bootstrap grids does in this case seem to only get in the
way. Plain percent would perhaps work better.)