docs: split vcs_support into admin/vcs_setup and usage/vcs_notes
The existing page on VCS support was a mix of information needed to setup
Kallithea with respect to version control systems, with information
regarding using version control systems (or specific aspects of it) with
Kallithea.

Move the first part to the Administrator Guide, and rebrand the second part
as VCS Usage Notes.

In vcs_notes.rst, the general info is moved above the Mercurial-specific
part, but otherwise left untouched.