ishell: specify ipython minimum version 4
`gearbox ishell` has an "optional" dependency on ipython as it fails with:
  Kallithea ishell requires the IPython Python package
Before the previous change, after installing ipython < 4, it would however
still fail the same way.
With the previous change, it would fail in a slightly more helpful way with:
  ImportError: No module named traitlets.config.loader

With ipython < 4, after installing the missing traitlets (4.3.2) as hinted,
ishell still failed with:
  TraitError: The 'config' trait of an InteractiveShellEmbed instance must be a Config or None, but a value of class 'traitlets.config.loader.Config' (i.e. {'InteractiveShellEmbed': {'confirm_exit': False}}) was specified.

With ipython >= 4, traitlets is installed as dependency, and ishell works.
Tested with both ipython 4.0 and current latest version 5.8 .

Thus, just clarify that we only support ipython >= 4 .