style: remove vertical scrollbar on hover of last line's comment bubble
When hovering the comment bubble of the last line of a diff, a vertical
scrollbar appears. This behavior is not present on the stable branch.

Attempts to fix the problem by changing element sizes or padding did not
solve the problem. As an alternative, this commit hides overflow - this makes
no visual difference.