front-end: Move .less files to the front-end folder
The less directory was public - we prefer to have the source file outside the
public folder. And we prefer to name the folder after the use of content
instead of the primary file type of content.

The last remaining 3rd-party file, mergely.css, was still not used from that
location and is just removed.