middleware: remove access fallback to reuse previous access - drop _git_stored_op
Before, the previous action was kept in the global controller instance. That
was conceptually wrong. The previous request might be entirely unrelated,
coming from another user.

It was mainly used for 'info/refs' commands ... but even more, that will be the
first command that is sent, giving nothing relevant to reuse.

Fortunately, with handling of 'info/refs', we no longer seem to need it.

The fallback for unknown commands with unknown 'action' is now to return a HTTP
failure, like we do for Mercurial.