setup: restrict TurboGears2 version to 2.3.x
In TurboGears2 2.3.12, the latest version in the 2.3.x range, the WebOb
dependency requirement is [1]
    WebOb >= 1.2, < 1.8.0

In TurboGears2 2.4.0 (which is in pre-release state at the time of this
commit), this becomes [2]:
    WebOb >= 1.8.0, < 1.10.0

In the Kallithea dependency list, we have matched our WebOb version
requirements to that of TurboGears2 and use:
    WebOb >= 1.7, < 1.8

while our TurboGears2 requirement was liberal and accepted anything in the
2.x range:
    TurboGears2 >= 2.3.10, < 3

To avoid new Kallithea installations failing with conflicting WebOb version
requirements after TurboGears2 2.4.0 is released, restrict the version of
TurboGears2 to 2.3.x on the stable branch.

For the default branch, the update to TurboGears2 2.4.0 can be considered
once it's released.

[1] https://github.com/TurboGears/tg2/blob/tg2.3.12/setup.py#L54
[2] https://github.com/TurboGears/tg2/blob/ed89788c3f5dab5a182a938543c9ee4ec14dd7ef/setup.py#L41