docs: move Apache+mod_wsgi example code to the corresponding bullets
The documentation about Apache+mod_wsgi has bullet points with inline
snippets, yet the example WSGI dispatch script is placed at the bottom of
the section instead of near its corresponding bullet.

It seems more readable and more according to the logical setup flow to move
the code next to its bullet.

Due to the additional indentation required to 'attach' the code to the
bullet, this commit is best viewed with the 'ignore whitespace changes'
setting.