tests: run doctests via pytest
pytest can run doctests as part of the standard testsuite run. See
documentation at [1].

pytest will discover doctests in all python files it can find. However, some
files cannot be imported directly in this manner.
Fix this by adding a filter in conftest.py in the root directory. This code
could also live in kallithea/conftest.py, but it cannot be in
kallithea/tests/conftest.py because that level is deeper than the files we
want to filter.

[1] https://docs.pytest.org/en/latest/doctest.html