i18n: remove fuzzy markers
Remove fuzzy markers in translation files.

This information can be useful to translators in the future and therefore
will be preserved in a separate kallithea-i18n repository with separate
*-i18n branches.

This commit is part of a series that normalizes the kallithea.po files by
removing all comments, in particular source code references.