i18n: update name/email of 'marcinkuzminski'
In scripts/contributor_data.py, there are already two 'name fixes' for
'marcink' and 'Marcin Kuzminski' to align all contributions with these names
to a specific name and email address.

Assuming they want to use the single email address as specified earlier, we
map the entry 'marcinkuzminski' in the i18n .po files on the same
name/email.