git: exit early if Git is found but too old (Issue #342)
The warning in the logs were too easy to miss.

It is still OK if git isn't found at all... for example if git_path
configuration doesn't point at something that looks like git.

The git_path configuration already had no default, so just make the code path
for that case more clean. (An easy next/alternative step could be to remove git
from BACKENDS if it isn't configured ...)

The system exit is similar to what is done in 0e33880b2897 ... even though
exiting from from a library might be a bit worse ...