i18n: drop automaticly introduced bad translations - completely remove previously fuzzy translations, not just the fuzzy markers
This is redoing cb02ca192180, but more aggressively.

msgmerge is aggressively adding fuzzy translations of strings that vaguely
resemble each other. But these translations are bad. And even more bad almost
irreversible when we dropped the fuzzy markers.

Consider for example the translation in 0.4.1 at
https://kallithea-scm.org/repos/kallithea/files/da65398a62fff50f3d241796cbf17acdea2092ef/kallithea/i18n/be/LC_MESSAGES/kallithea.po

"Unauthorized access to resource" is translated to "Несанкцыянаваны доступ да
рэсурсу", but the same translation is also used fuzzily for very different
strings. That can't be good. While the translation of a slightly similar string
might be a good starting point for translators, we don't want to show such
automated and invalid translations to our users. Then it is better to show the
orginal english text, that at least is a clear TODO item and doesn't say
anything wrong.

We thus remove these automaticly generated translations. Some of them were
perhaps marked fuzzy for other reasons, and we might lose a few good
translations. That is still better than having wrong translations. Translators
can also just mark the translation as non-fuzzy and deal with the conflict of
diverging opinion on whether the translation is good. Their statement trumps
our presumption that the automated fuzzy translation is bad.