page: minimal change to move from webhelpers.paginate to paginate
webhelpers is dead and doesn't work with py3. paginate is not very actively
maintained, but it is the natural successor to webhelpers.paginate, it seems
stable, and it works with py3.

This is a minimal change that seems to work. It preserves existing tech debt
... and adds a little bit more. It will be cleaned up next.

webhelpers.paginate had built-in SqlAlchemy support - now we have to handle it
explicitly.