admin: fix 'Settings > Visual' form validation after commit 574218777086
Commit 574218777086 introduced a setting for 'SSH Clone URL' in 'Admin >
Settings > Visual' and placed it under a check 'if c.ssh_enabled', which
means that the corresponding form field is not present when SSH is not
enabled.
In this case, when trying to save the form (changing any or none setting),
form validation reports an error 'Missing value' without much detail. At the
top of the HTML document, even before the opening HTML tag, we can see:

    <!-- for: clone_ssh_tmpl -->
    <span class="error-message">Missing value</span><br />


Fix this problem by adding a hidden form field for clone_ssh_tmpl, with the
current value from the database, in case SSH is not enabled.