ssh: drop usk_public_key_idx again
Essentially a backout of d2a97f73fa1f and the
4851d15bc437_db_migration_step_after_95c01895c006_ alembic step.

We can't reliably have full index on fields with unbounded length. The
upgrade step has been reported to fail on MySQL [1]:

    sqlalchemy.exc.OperationalError: (_mysql_exceptions.OperationalError)
    (1170, "BLOB/TEXT column 'public_key' used in key specification without
    a key length") [SQL: u'CREATE INDEX usk_public_key_idx ON user_ssh_keys
    (public_key)'] (Background on this error at: http://sqlalche.me/e/e3q8)

And we really don't need this index ... especially now when we use
fingerprints for key deletion instead of looking up by the full public key.


[1] https://lists.sfconservancy.org/pipermail/kallithea-general/2019q4/003068.html